tester

+++ ft_strrchr.c
@@ -12,20 +12,20 @@
 
 #include "libft.h"
 
-char	*ft_strrchr(const char *s, int c)
+char *ft_strrchr(const char *s, int c)
 {
-	size_t	i;
+        int i;
 
-	if (!s)
-		return (NULL);
-	i = ft_strlen(s);
-	while (i >= 0)
-	{
-		if (s[i] == (char)c)
-			return ((char *)s + i);
-		i--;
-	}
-	return (NULL);
+        if (!s)
+                return (NULL);
+        i = ft_strlen(s);
+        while (i >= 0)
+        {
+                if (s[i] == (char)c)
+                        return ((char *)s + i);
+                i--;
+        }
+        return (NULL);
 }
 /*
 int     main(void)
